cau h, i

+++ dsHoaDon.h
@@ -29,12 +29,13 @@ struct ds_hoa_don {
 	int n_hd = 0;
 };
 
+
 void Init_hd(hoa_don& hd);
 hd_Node* MakeHdNode(hoa_don data);
 void AddLastListHD(ds_hoa_don*& dshd, hoa_don hd);
 void InputFormHD(string ct[], int sl, int length);
 hd_Node* FindSoHDExist(ds_hoa_don* dshd, string ID);
-double ComputeValue(int sl, double donGia, double VAT);
+int ComputeValue(int sl, double donGia, double VAT);
 int PickHDType();
 void ButtonFunction();
 void TableOfHD();
